+++ tests/test-vim.sh
@@ -2,4 +2,4 @@
 echo "Installing Vim plugins..."
 vim +PlugInstall +qall!
 echo "Vim exited."
-ps aux | grep -v grep | grep -E "vim|go"
+ps aux | grep -v grep | grep -E "vim|go" || true